move creating a sitemap to build script
