+++ .gitignore
@@ -38,3 +38,5 @@ yarn-error.log*
 
 .yarn
 yarn.lock
+
+public/sitemap.xml
