+++ ci/tsconfig.json
@@ -13,6 +13,6 @@
         "rootDir": "./",
         "outDir": "../dist/ci"
   },
-  "include": ["**/*.ts"],
+  "include": ["**/*.ts", "sitemap.js"],
   "exclude": ["node_modules"]
 }
