+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "build": "pnpm run sitemap && next build",
     "build:ci": "pnpm run build && pnpm run build:cionly",
     "build:cionly": "tsc --build ci",
     "start": "next start",
@@ -13,7 +13,8 @@
     "lint:fix": "eslint components contexts hooks pages --fix",
     "prettier": "prettier -c **/*.mdx **/*.ts{,x}",
     "prettier:fix": "prettier --write **/*.mdx **/*.ts{,x}",
-    "test:links": "node dist/ci/checkLinks.js"
+    "test:links": "node dist/ci/checkLinks.js",
+    "sitemap": "node ci/sitemap.js"
   },
   "dependencies": {
     "@docsearch/react": "^3.6.0",
