+++ tsconfig.json
@@ -23,7 +23,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "ci/sitemap.js"  ],
   "exclude": [
     "node_modules"
   ]
